Add files via upload
This is a test file of a photoactivation experiment. Notice that its header contains Time, Time.1, [s], Intensity, etc. which are quite fragmented. When convert to csv file, these headers will not make sense. So this python script removes the header all together to allow the user to customisably add a new header in with their desire titles.